Tests/LibWeb/Text/expected/css/stylesheet-add-remove-column-trailing-universal-invalidation-counters.txt

column trailing-universal add invalidations: 5
